redo.py: don't add dependencies on files we rebuild.
redo-ifchange already does this, and then calls redo.py, so it's redundant.

The only weird case is if a .do file calls 'redo' instead of
'redo-ifchange'.  I'm not sure what exactly is supposed to happen here.
test.do does this, for example, and it works, but only because we always do
'redo test' and nobody depends on 'redo-ifchange test'; if they did, the
dependencies would be off, I guess.  But maybe it's meaningless to do that
anyway.