jwack: start waitfds around fd#50.
That makes it a little easier to tell, in a strace, what the process is
waiting on.  If it's 100/101, then it's waiting on a token; 50+ means waiting
on a subtask.

Also, we weren't closing the read side of subtask fds on exec.  This didn't
cause any problems, but did result in a wasted fd in subprocesses.