Fix assertion error from double-lock of target
When a target appears twice in the same call but with slightly different names
(perhaps one has ./ in front) then the old string comparison of the target
names wasn't enough to prevent the target from being attempted twice.

Convert the de-duping logic to use the file-id which ensures that the target
will not be attempted twice -- even if it has a slightly different name.