Change license to Apache 2.0.
Mailing list discussion was here:
https://groups.google.com/forum/#!topic/redo-list/wLMZMxtn4wo

Several more contributors replied to me personally to say that they
don't have a problem with the change (and several consider the change
to be an improvement).

The overwhelming majority of everything in the redo repo was written by
me, so I have the right to change the license unilaterally anyway,
subject to a few rules.  Since the new license actually removes
licensing/usage restrictions for everyone, this should be no problem.