sh.do: explicitly check /bin/sh instead of sh.
Because our experiment involves creating a file called redo/sh, and the
redo directory (or some other version of redo's dirctory) might be on
the path, just testing 'sh' is actually wrong: we might end up with
some earlier version of redo-sh.

Instead, specifically for sh, always demand that we test /bin/sh.